code/Tools/RZCameraExtrinsicsTool/pyproject.toml
[project]
name = "RZCameraExtrinsicsTool"
version = "0.1.0"
authors = [
  { name="Peter Preinesberger", email="peter.preinesberger@study.thws.de" },
]
description = "A tool for visually approximating camera extrinsics."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux"
]
license-files = ["LICEN[CS]E*"]
dependencies = [
  "RZVisuals",
  "numpy",
  "Pillow"
]

[project.scripts]
rzcameraextrinsics = "RZCameraExtrinsicsTool.__main__:main"


[project.urls]
Homepage = "https://github.com/thws-mai/PM25_Mice"
Issues = "https://github.com/thws-mai/PM25_Mice/issues"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
